Add rayon-based parallelisation
