+++ Cargo.toml
@@ -48,6 +48,7 @@ tree-sitter-rust = "0.21"
 uniquote = "4.0.0"
 textwrap = { version = "0.16.1", default-features = false }
 walkdir = "2"
+rayon = "1.10.0"
 
 [dev-dependencies]
 insta = { version = "1.36.1", features = ["yaml"] }
