+++ src/cli.rs
@@ -167,7 +167,7 @@ impl Display for MaxConcurrentFileLimit {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum MaxProblems {
     Unlimited,
     Limited(u32),
